Refactor folders : solve #4
