+++ cpmac/app/drawers/cisc/cisc-page.js
@@ -5,7 +5,7 @@ const FileIO = require("../../shared/lib/file-io/file-io");
 const ObservableModule = require("data/observable");
 const ObservableArray = require("data/observable-array").ObservableArray;
 const BarcodeScanner = require("nativescript-barcodescanner").BarcodeScanner;
-const CothorityMessages = require("~/shared/lib/dedjs/protobuf/build/cothority-messages");
+const CothorityMessages = require("~/shared/lib/dedjs/network/cothority-messages");
 const User = require("~/shared/lib/dedjs/object/user/User").get;
 const Dialog = require("ui/dialogs");
 const Helper = require("~/shared/lib/dedjs/Helper");
