+++ cpmac/app/drawers/cisc/data/data-page.js
@@ -1,7 +1,7 @@
 const FrameModule = require("ui/frame");
 const Dialog = require("ui/dialogs");
 const RequestPath = require("~/shared/lib/dedjs/RequestPath");
-const CothorityMessages = require("~/shared/lib/dedjs/protobuf/build/cothority-messages");
+const CothorityMessages = require("~/shared/lib/dedjs/network/cothority-messages");
 const DecodeType = require("~/shared/lib/dedjs/DecodeType");
 const DedisJsNet = require("~/shared/lib/dedjs/Net");
 const Convert = require("~/shared/lib/dedjs/Convert");
