+++ cpmac/app/drawers/cisc/home/home-page.js
@@ -1,19 +1,5 @@
 const FrameModule = require("ui/frame");
 const Dialog = require("ui/dialogs");
-const RequestPath = require("~/shared/lib/dedjs/RequestPath");
-const CothorityMessages = require("~/shared/lib/dedjs/protobuf/build/cothority-messages");
-const DecodeType = require("~/shared/lib/dedjs/DecodeType");
-const DedisJsNet = require("~/shared/lib/dedjs/Net");
-const Convert = require("~/shared/lib/dedjs/Convert");
-const BarcodeScanner = require("nativescript-barcodescanner").BarcodeScanner;
-const Helper = require("~/shared/lib/dedjs/Helper");
-const FileIO = require("../../../shared/lib/file-io/file-io");
-const Cisc = require("~/shared/lib/dedjs/object/cisc/Cisc").Skipchain;
-const User = require("~/shared/lib/dedjs/object/user/User").get;
-const NetDedis = require("@dedis/cothority").net;
-const ObservableModule = require("data/observable");
-const ObservableArray = require("data/observable-array").ObservableArray;
-const FilePaths = require("../../../shared/res/files/files-path");
 const SkipPage = require("../skipchain-page");
 
 let page = undefined;
