+++ cpmac/app/drawers/home/home-page.js
@@ -3,7 +3,7 @@ const Dialog = require("ui/dialogs");
 const Helper = require("../../shared/lib/dedjs/Helper");
 const Convert = require("../../shared/lib/dedjs/Convert");
 const ObjectType = require("../../shared/lib/dedjs/ObjectType");
-const CothorityMessages = require("../../shared/lib/dedjs/protobuf/build/cothority-messages");
+const CothorityMessages = require("../../shared/lib/dedjs/network/cothority-messages");
 const ScanToReturn = require("../../shared/lib/scan-to-return/scan-to-return");
 const User = require("../../shared/lib/dedjs/object/user/User").get;
 const ObservableModule = require("data/observable");
