+++ cpmac/app/drawers/pop/org/config/config-page.js
@@ -2,7 +2,6 @@ const Dialog = require("ui/dialogs");
 const Frame = require("ui/frame");
 const Helper = require("../../../../shared/lib/dedjs/Helper");
 const Convert = require("../../../../shared/lib/dedjs/Convert");
-const Net = require("../../../../shared/lib/dedjs/Net");
 const ObjectType = require("../../../../shared/lib/dedjs/ObjectType");
 const ScanToReturn = require("../../../../shared/lib/scan-to-return/scan-to-return");
 const Observable = require("tns-core-modules/data/observable");
@@ -355,79 +354,6 @@ function hashAndSave() {
   return registerPopDesc();
 }
 
-function manageDesc() {
-  return Dialog.confirm({
-    title: "PoP-Description",
-    message: "Do you want to share your description or import a new one?",
-    okButtonText: "Import",
-    cancelButtonText: "Cancel",
-    neutralButtonText: "Share"
-  })
-    .then(result => {
-      if (result) {
-        // Import
-        return ScanToReturn.scan()
-          .then(pasteBinIdJson => {
-            const id = Convert.jsonToObject(pasteBinIdJson).id;
-            const PasteBin = new Net.PasteBin();
-
-            return PasteBin.get(id);
-          })
-          .then(popDescJson => {
-            const popDesc = Convert.parseJsonPopDesc(popDescJson);
-
-            return Party.setPopDesc(popDesc, true)
-              .then(() => {
-                return setUpDate();
-              });
-          })
-      } else if (result === undefined) {
-        // Share
-        if (!Party.isPopDescComplete()) {
-          return Dialog.alert({
-            title: "Missing Information",
-            message: "Please provide a name, date, time, location and the list (min 3) of conodes" +
-            " of the organizers of your PoP Party.",
-            okButtonText: "Ok"
-          });
-        }
-
-        const PasteBin = new Net.PasteBin();
-        const popDescJson = JSON.stringify(Party.getPopDesc());
-
-        return PasteBin.paste(popDescJson)
-          .then(id => {
-            const object = {};
-            object.id = id;
-
-            pageObject.showModal("shared/pages/qr-code/qr-code-page", {
-              textToShow: Convert.objectToJson(object),
-              title: "Party informations"
-            }, () => {
-            }, true);
-
-            return Promise.resolve();
-          });
-      } else {
-        // Cancel
-        return Promise.resolve();
-      }
-    })
-    .catch(error => {
-      console.log(error);
-      console.dir(error);
-      console.trace();
-
-      Dialog.alert({
-        title: "Error",
-        message: "An error occured, please try again. - " + error,
-        okButtonText: "Ok"
-      });
-
-      return Promise.reject(error);
-    });
-}
-
 function goBack() {
   topmost().goBack();
 }
@@ -501,7 +427,6 @@ module.exports.addManual = addManual;
 module.exports.addScan = addScan;
 module.exports.deleteConode = deleteConode;
 module.exports.onSwipeCellStarted = onSwipeCellStarted;
-module.exports.manageDesc = manageDesc;
 module.exports.goBack = goBack;
 module.exports.finish = finish;
 module.exports.addOrganizer = addOrganizer;
