+++ cpmac/app/shared/lib/dedjs/network/DecodeType.js
