+++ cpmac/app/shared/lib/dedjs/object/cisc/Cisc.js
@@ -9,11 +9,11 @@ const ObjectType = require("../../ObjectType");
 const Helper = require("../../Helper");
 const Convert = require("../../Convert");
 const FilesPath = require("../../../../res/files/files-path");
-const DecodeType = require("../../DecodeType");
+const DecodeType = require("../../network/DecodeType");
 const FileIO = require("../../../../lib/file-io/file-io");
 const Crypto = require("../../Crypto");
-const RequestPath = require("../../RequestPath");
-const CothorityMessages = require("../../protobuf/build/cothority-messages");
+const RequestPath = require("../../network/RequestPath");
+const CothorityMessages = require("../../network/cothority-messages");
 const BigNumber = require("bn.js");
 const User = require("../user/User").get;
 const FrameModule = require("ui/frame");
