+++ cpmac/app/shared/lib/dedjs/object/user/User.js
@@ -12,9 +12,9 @@ const ObjectType = require("../../ObjectType");
 const Helper = require("../../Helper");
 const NetDedis = require("@dedis/cothority").net;
 const Crypto = require("../../Crypto");
-const RequestPath = require("../../RequestPath");
-const DecodeType = require("../../DecodeType");
-const CothorityMessages = require("../../protobuf/build/cothority-messages");
+const RequestPath = require("../../network/RequestPath");
+const DecodeType = require("../../network/DecodeType");
+const CothorityMessages = require("../../network/cothority-messages");
 
 /**
  * This singleton is the user of the app. It contains everything needed that is general, app-wide or does not belong to any precise subpart.
