+++ cpmac/app/tests/lib/dedjs/object/pop/PoPTest.js
@@ -5,7 +5,7 @@ chai.should();
 const FilesPath = require("../../../../../shared/res/files/files-path");
 const FileIO = require("../../../../../shared/lib/file-io/file-io");
 const Convert = require("../../../../../shared/lib/dedjs/Convert");
-const CothorityMessages = require("../../../../../shared/lib/dedjs/protobuf/build/cothority-messages");
+const CothorityMessages = require("../../../../../shared/lib/dedjs/network/cothority-messages");
 const User = require("../../../../../shared/lib/dedjs/object/user/User").get;
 
 const PoP = require("../../../../../shared/lib/dedjs/object/pop/PoP").get;
