+++ cpmac/app/tests/lib/dedjs/object/user/UserTest.js
@@ -12,7 +12,7 @@ const Convert = require("../../../../../shared/lib/dedjs/Convert");
 const Helper = require("../../../../../shared/lib/dedjs/Helper");
 const ObjectType = require("../../../../../shared/lib/dedjs/ObjectType");
 const Crypto = require("../../../../../shared/lib/dedjs/Crypto");
-const CothorityMessages = require("../../../../../shared/lib/dedjs/protobuf/build/cothority-messages");
+const CothorityMessages = require("../../../../../shared/lib/dedjs/network/cothority-messages");
 
 const User = require("../../../../../shared/lib/dedjs/object/user/User").get;
